Add support for building container images from pbr source
All OpenStack pbr packages can be turned in to single-process docker
containers the same way. bindep files describe the distro-packages
needed. requirements.txt describe the requirements and setup.cfg lists
the programs that get installed.

Occasionally, like with zuul-executor from Zuul, there are depends that
don't quite fit the mold, so allow a final-stage Dockerfile override.

This uses python:slim as a base image and does not make that
configurable. That is because these are python programs, so making
containers for different base distros doesn't make any sense.