Switch from python:slim to python:alpine
Alpine is smaller. Also, for our first user, Zuul, alpine already has
all the things needed from extra repos (yarn and bubblewrap) Switch to
alpine so we can keep things simple.

NOTE: This isn't going to work until we cut a new release of bindep with
the depends-on patch in it.

Depends-On: https://review.openstack.org/579056
Change-Id: I8c8a3b94b5aa562f54034825c3af871893b4bf89