Fix the alpine build
Ok. So the previous alpine patch didn't work QUITE as much as you might
want.

In fact, it didn't work even a little bit.

Update it to work.

But also, change it a little bit and allow for expressing per-image
additional dependencies in python extras. If there is an extra matching
the image name, it'll get installed into that image.

Change-Id: I9eda25829be6d2916bb131cb6bc8f0382ef44f1b