Clear up "tag" vs "image"
We incorrectly reference things in the code as "tags" when what we
really mean is "image". We are only using actual tagging to create
a new prefixed image name that references the original image.

This changes code references from "tag" to "image" in places it
makes sense. Also, we no longer use tagging to create the prefixed
image name, and instead only create the prefixed image and commit
over it, leaving the resulting local images looking much cleaner
(with no duplicates).

Change-Id: I6234a62d927cb9f35f8f4b731eee375f4c208870