Run an sdist before the bdist_wheel command
The hook sequencing doesn't work as well for bdist_wheel as it does for
sdist because bdist_wheel creates a build destination before setup hooks
run, but the setup hooks get run the context of the source dir, so that
the content gets generated but not in the right location.

sdist does not do a build directory, so by running an sdist first,
we cause setup hooks to get run in the correct context and before the
bdist_wheel step gets run. This means when bdist_wheel gets run, the
results of any setup hooks will be appropriately in place.

Change-Id: I5ca72b9cebcb27d7ab7b44198a98f7a8287de0b3